Test Unit cases that I'm going to create first.
ASync Task test cases are a bit tougher, so I may look into a few of the other forms of testing.

Doesn't make a ton of sense to make test cases for Objects, so I created some for their helpers.

Will be updating with code over the course of the evening